LayoutTests/imported/w3c/web-platform-tests/html/browsers/offline/appcache/workers/appcache-worker.https-expected.txt

CONSOLE MESSAGE: ApplicationCache is deprecated. Please use ServiceWorkers instead.

Harness Error (TIMEOUT), message = null

PASS Dedicated worker of the cached script
PASS Dedicated worker of the fallbacked script
PASS Dedicated worker of the not-in-cache script
FAIL Shared worker of the cached script assert_equals: expected "Done: cached" but got "Error: Importing a non-cached script must fail."
FAIL Shared worker of the fallbacked script assert_equals: expected "Done: fallbacked" but got "Error: Importing a non-cached script must fail."
TIMEOUT Shared worker of the not-in-cache script Test timed out
